👌 ADD: Introduce PlayerState interface and enhance player messaging; refactor video handling in Playlist component
