+++ background-script/server.ts
@@ -1,11 +1,12 @@
 import { BackgroundMessagingClient } from "./clients/ExtensionClient";
 import ServerClient from "./clients/ServerClient";
-import { notifyTabsPrimaryTabSet, setPrimaryTab } from "./tab";
+import { setPrimaryTab } from "./tab";
 import { ExtensionMessageType } from "types/extensionMessage";
 import { profile } from "types/profile";
 import {
     FromServerMessageType,
     Member,
+    Player,
     Playlist,
     ToServerMessageType,
     Video,
@@ -15,6 +16,7 @@ const server = ServerClient.getInstance();
 const message = BackgroundMessagingClient.getInstance();
 
 interface State {
+    player: Player;
     room_id: string;
     members: Member[];
     playlist: Playlist;
@@ -22,22 +24,23 @@ interface State {
 }
 
 const state: State = {
+    player: {} as Player,
     room_id: "",
     members: [] as Member[],
     playlist: {
         videos: [] as Video[],
-        last_video_id: null,
+        last_video: null,
     },
     is_admin: false,
 };
 
 server.addHandler(FromServerMessageType.JOINED_ROOM, payload => {
+    state.player = payload.room.player;
     state.playlist = payload.room.playlist;
     state.members = payload.room.members;
     state.room_id = payload.room.room_id;
     state.is_admin = payload.joined_member.is_admin;
-    const video_url = payload.room.player.video_url;
-    chrome.tabs.create({ url: `https://youtube.com/watch?v=${video_url}` }, tab => {
+    chrome.tabs.create({ url: `https://youtube.com/watch?v=${state.player.video_url}` }, tab => {
         if (tab.id) setPrimaryTab(tab.id);
     });
 });
@@ -84,6 +87,22 @@ export const updateProfile = (profile: profile) => {
     server.send(ToServerMessageType.UPDATE_PROFILE, profile);
 };
 
+server.addHandler(FromServerMessageType.PLAYER_VIDEO_UPDATED, payload => {
+    state.playlist = payload.playlist;
+    state.player = payload.player;
+    message.sendMessageToPrimaryTab(ExtensionMessageType.PLAYLIST_UPDATED, payload.playlist);
+    message.sendMessageToPrimaryTab(
+        ExtensionMessageType.PLAYER_VIDEO_UPDATED,
+        payload.player.video_url,
+    );
+    if (payload.playlist.last_video)
+        message.sendMessageToPrimaryTab(
+            ExtensionMessageType.PREVIOUS_VIDEO_UPDATED,
+            payload.playlist.last_video,
+        );
+    message.sendMessageToPrimaryTab(ExtensionMessageType.UPDATE_PLAYER_STATE, payload.player);
+});
+
 // Message
 
 message.addHandler(ExtensionMessageType.ADD_VIDEO, (videoUrl: string) => {
@@ -116,4 +135,23 @@ message.addHandler(ExtensionMessageType.PROMOTE_USER, id => {
 
 message.addHandler(ExtensionMessageType.REMOVE_MEMBER, id => {
     server.send(ToServerMessageType.REMOVE_MEMBER, { member_id: id });
+});
+
+message.addHandler(ExtensionMessageType.UPDATE_PLAYER_VIDEO, payload => {
+    server.send(ToServerMessageType.UPDATE_PLAYER_VIDEO, {
+        video_id: payload,
+        updated_at: Date.now(),
+    });
+});
+
+message.addHandler(ExtensionMessageType.GET_PLAYER_STATE, () => {
+    return state.player;
+});
+
+message.addHandler(ExtensionMessageType.GET_PLAYER_VIDEO, () => {
+    return state.player.video_url;
+});
+
+message.addHandler(ExtensionMessageType.GET_PREVIOUS_VIDEO, () => {
+    return state.playlist.last_video;
 });
